backend/requirements.txt
aiosmtplib==3.0.1
email-validator>=2.0.0
aiosqlite==0.19.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
black==24.1.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.1
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
flake8==7.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.29.0
google-auth==2.47.0
google-genai==1.60.0
google-generativeai==0.3.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==5.13.2
jiter==0.12.0
mccabe==0.7.0
mypy==1.8.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==2.15.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.9.12
packaging==26.0
pathspec==1.0.3
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
pyflakes==3.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tomli==2.4.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
